"Leave Management Endpoints"

+++ src/main/java/za/co/mawa/bes/dto/leave/request/LeaveRequestOutboundDto.java
@@ -19,6 +19,6 @@ public class LeaveRequestOutboundDto {
     private PartnerDto approver;
     private Date startDate;
     private Date endDate;
-    private int days;
+    private long days;
     private FieldOptionDto status;
 }
